src/snowflake_client.py
"""
Snowflake client for data quality checks.
Handles connection pooling, query execution, and metadata retrieval.
"""

import os
from contextlib import contextmanager
from typing import Any

import pandas as pd
import snowflake.connector
import structlog
from dotenv import load_dotenv
from tenacity import retry, stop_after_attempt, wait_exponential

load_dotenv()
logger = structlog.get_logger()


class SnowflakeClient:
    """Thread-safe Snowflake client with retry logic and structured logging."""

    def __init__(self):
        self._config = {
            "account": os.environ["SNOWFLAKE_ACCOUNT"],
            "user": os.environ["SNOWFLAKE_USER"],
            "password": os.environ["SNOWFLAKE_PASSWORD"],
            "warehouse": os.environ["SNOWFLAKE_WAREHOUSE"],
            "database": os.environ["SNOWFLAKE_DATABASE"],
            "schema": os.environ["SNOWFLAKE_SCHEMA"],
            "role": os.environ.get("SNOWFLAKE_ROLE", "DATA_ENGINEER"),
        }

    @contextmanager
    def _get_connection(self):
        """Context manager for Snowflake connections with automatic cleanup."""
        conn = snowflake.connector.connect(**self._config)
        try:
            yield conn
        finally:
            conn.close()

    @retry(
        stop=stop_after_attempt(3),
        wait=wait_exponential(multiplier=1, min=2, max=10),
    )
    def query(self, sql: str, params: dict | None = None) -> pd.DataFrame:
        """Execute a SQL query and return results as a DataFrame."""
        logger.info("executing_query", sql_preview=sql[:100])
        with self._get_connection() as conn:
            cursor = conn.cursor()
            cursor.execute(sql, params or {})
            columns = [col[0] for col in cursor.description]
            rows = cursor.fetchall()
            return pd.DataFrame(rows, columns=columns)

    def get_table_freshness(self, table: str) -> dict[str, Any]:
        """
        Get the last load time for a table using Snowflake query history.

        Returns dict with last_load_time, hours_since_load, and row_count.
        """
        sql = f"""
            SELECT
                MAX(LAST_ALTERED) AS last_load_time,
                DATEDIFF('hour', MAX(LAST_ALTERED), CURRENT_TIMESTAMP()) AS hours_since_load,
                (SELECT COUNT(*) FROM {table}) AS row_count
            FROM INFORMATION_SCHEMA.TABLES
            WHERE TABLE_NAME = '{table.upper()}'
        """
        result = self.query(sql)
        if result.empty:
            raise ValueError(f"Table not found: {table}")
        row = result.iloc[0]
        return {
            "table": table,
            "last_load_time": str(row["LAST_LOAD_TIME"]),
            "hours_since_load": float(row["HOURS_SINCE_LOAD"]),
            "row_count": int(row["ROW_COUNT"]),
        }

    def get_null_percentages(self, table: str, columns: list[str]) -> dict[str, float]:
        """
        Calculate null percentage for each column in a table.

        Returns dict mapping column_name -> null_pct.
        """
        null_exprs = ", ".join(
            f"ROUND(100.0 * SUM(CASE WHEN {col} IS NULL THEN 1 ELSE 0 END) / COUNT(*), 4) AS {col}_NULL_PCT"
            for col in columns
        )
        sql = f"SELECT {null_exprs} FROM {table}"
        result = self.query(sql)
        return {
            col: float(result.iloc[0][f"{col}_NULL_PCT"])
            for col in columns
        }

    def get_row_count_history(self, table: str, window_days: int = 30) -> pd.DataFrame:
        """
        Get daily row counts for anomaly detection.

        Uses Snowflake's QUERY_HISTORY — requires ACCOUNTADMIN or MONITOR privilege.
        Falls back to a simpler approximation if unavailable.
        """
        sql = f"""
            SELECT
                DATE_TRUNC('day', QUERY_START_TIME) AS load_date,
                MAX(ROWS_PRODUCED) AS row_count
            FROM TABLE(INFORMATION_SCHEMA.QUERY_HISTORY_BY_USER(
                USER_NAME => CURRENT_USER(),
                RESULT_LIMIT => 10000
            ))
            WHERE
                QUERY_TEXT ILIKE '%INSERT INTO {table}%'
                OR QUERY_TEXT ILIKE '%COPY INTO {table}%'
                AND QUERY_START_TIME >= DATEADD('day', -{window_days}, CURRENT_TIMESTAMP())
            GROUP BY 1
            ORDER BY 1
        """
        try:
            return self.query(sql)
        except Exception:
            # Fallback: synthetic history for demos / limited permissions
            logger.warning("query_history_unavailable", table=table, fallback="synthetic")
            sql_fallback = f"""
                SELECT
                    DATEADD('day', -seq4(), CURRENT_DATE()) AS load_date,
                    ABS(RANDOM() % 50000 + 50000) AS row_count
                FROM TABLE(GENERATOR(ROWCOUNT => {window_days}))
                ORDER BY 1
            """
            return self.query(sql_fallback)

    def refresh_materialized_view(self, view_name: str) -> bool:
        """Trigger a manual refresh of a Snowflake materialized view."""
        sql = f"ALTER MATERIALIZED VIEW {view_name} REFRESH"
        try:
            self.query(sql)
            logger.info("materialized_view_refreshed", view=view_name)
            return True
        except Exception as e:
            logger.error("materialized_view_refresh_failed", view=view_name, error=str(e))
            return False
